Detection/Software/pyproject.toml
[tool.poetry]
name = "software"
version = "0.1.1"
description = ""
authors = ["Rodrigo <rodr@itu.dk>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
fastapi = "^0.97.0"
uvicorn = {extras = ["standard"], version = "^0.22.0"}
numpy = "^1.24.3"
asyncio = "^3.4.3"
pandas = "^2.0.2"
scipy = "^1.10.1"
pycmdmessenger = "^0.2.4"
pyserial = "^3.5"
opencv-python = "^4.7.0.72"

[tool.poetry.scripts]
lle-fastapi = "software.API.FastAPI.BeltDrainLLE.BeltDrainLLE:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
